Importing genius yet again after a period of not working on it. I really like it for homework far better then octave ... it can even output troff/latex code for matrixes for me.  Missing a lot of ChangeLog because I was lazy.
This is the stuff that was on the DrGenius work without all the CLN and
multiple execution context rewrite crap which I would never have time to finish
anyway.  So it's not as beautiful as it should have been, but who cares.  It
works.  Mostly anyway.

I changed the syntax too, just for good meassure.  And it still uses gnome1 stuff
because I'm lazy / don't have time

-George